+++ otter/assign/__init__.py
@@ -50,17 +50,14 @@ def main(master, result, *, no_pdfs=False, no_run_tests=False, username=None, pa
         # update seed variables
         if assignment.seed.variable:
             LOGGER.debug("Processing seed dict")
-            if assignment.generate_enabled:
-                LOGGER.debug("Otter Generate configuration found while processing seed dict")
-                assignment.generate.seed = assignment.seed.autograder_value
-                assignment.generate.seed_variable = assignment.seed.variable
-                LOGGER.debug("Added seed information to assignment.generate")
+            assignment.generate.seed = assignment.seed.autograder_value
+            assignment.generate.seed_variable = assignment.seed.variable
+            LOGGER.debug("Added seed information to assignment.generate")
 
         # check that we have a seed if needed
         if assignment.seed_required:
             LOGGER.debug("Assignment seed is required")
-            if assignment.generate_enabled and \
-                    not isinstance(assignment.generate.seed, int):
+            if not isinstance(assignment.generate.seed, int):
                 raise RuntimeError("Seeding cell found but no or invalid seed provided")
 
         plugins, pc = assignment.plugins, None
@@ -68,14 +65,12 @@ def main(master, result, *, no_pdfs=False, no_run_tests=False, username=None, pa
             LOGGER.debug("Processing plugins")
             pc = PluginCollection(plugins, "", {})
             pc.run("during_assign", assignment)
-            if assignment.generate_enabled:
-                LOGGER.debug("Adding plugin configurations to Otter Generate configuration")
-                assignment.generate.plugins.extend(plugins)
 
-        # generate Gradescope autograder zipfile
-        if assignment.generate_enabled:
-            LOGGER.info("Generating autograder zipfile")
-            run_generate_autograder(assignment, username, password, plugin_collection=pc)
+            LOGGER.debug("Adding plugin configurations to Otter Generate configuration")
+            assignment.generate.plugins.extend(plugins)
+
+        LOGGER.info("Generating autograder zipfile")
+        run_generate_autograder(assignment, username, password, plugin_collection=pc)
 
         # generate PDF of solutions
         if assignment.solutions_pdf and not no_pdfs:
@@ -151,17 +146,12 @@ def main(master, result, *, no_pdfs=False, no_run_tests=False, username=None, pa
         if assignment.run_tests and not no_run_tests and assignment.is_python:
             LOGGER.info("Running tests against the solutions notebook")
 
-            seed = assignment.generate.seed if assignment.generate_enabled else None
+            seed = assignment.generate.seed
             LOGGER.debug(f"Resolved seed for running tests: {seed}")
 
-            if assignment.generate_enabled:
-                LOGGER.debug("Retrieving updated plugins from otter_config.json for running tests")
-                test_pc = PluginCollection(
-                    assignment.generate.plugins, assignment.ag_notebook_path, {})
-
-            else:
-                LOGGER.debug("Using pre-configured plugins for running tests")
-                test_pc = pc
+            LOGGER.debug("Retrieving updated plugins from otter_config.json for running tests")
+            test_pc = PluginCollection(
+                assignment.generate.plugins, assignment.ag_notebook_path, {})
 
             run_tests(
                 assignment.get_ag_path(master.name),